Rename profiler to profilers (#12308)
