+++ examples/pl_basics/profiler_example.py
@@ -31,7 +31,7 @@ import torchvision.models as models
 import torchvision.transforms as T
 
 from pytorch_lightning import cli_lightning_logo, LightningDataModule, LightningModule
-from pytorch_lightning.profiler.pytorch import PyTorchProfiler
+from pytorch_lightning.profilers.pytorch import PyTorchProfiler
 from pytorch_lightning.utilities.cli import LightningCLI
 
 DEFAULT_CMD_LINE = (
