+++ src/pytorch_lightning/profilers/__init__.py
@@ -11,12 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from pytorch_lightning.profiler.advanced import AdvancedProfiler
-from pytorch_lightning.profiler.base import AbstractProfiler, BaseProfiler, PassThroughProfiler
-from pytorch_lightning.profiler.profiler import Profiler
-from pytorch_lightning.profiler.pytorch import PyTorchProfiler
-from pytorch_lightning.profiler.simple import SimpleProfiler
-from pytorch_lightning.profiler.xla import XLAProfiler
+from pytorch_lightning.profilers.advanced import AdvancedProfiler
+from pytorch_lightning.profilers.base import AbstractProfiler, BaseProfiler, PassThroughProfiler
+from pytorch_lightning.profilers.profiler import Profiler
+from pytorch_lightning.profilers.pytorch import PyTorchProfiler
+from pytorch_lightning.profilers.simple import SimpleProfiler
+from pytorch_lightning.profilers.xla import XLAProfiler
 
 __all__ = [
     "AbstractProfiler",
