+++ src/pytorch_lightning/profilers/base.py
@@ -15,7 +15,7 @@
 from abc import ABC, abstractmethod
 from typing import Any
 
-from pytorch_lightning.profiler.profiler import Profiler
+from pytorch_lightning.profilers.profiler import Profiler
 from pytorch_lightning.utilities.rank_zero import rank_zero_deprecation
