+++ src/pytorch_lightning/trainer/trainer.py
@@ -55,7 +55,7 @@ from pytorch_lightning.plugins import (
     PrecisionPlugin,
 )
 from pytorch_lightning.plugins.environments.slurm_environment import SLURMEnvironment
-from pytorch_lightning.profiler import (
+from pytorch_lightning.profilers import (
     AdvancedProfiler,
     PassThroughProfiler,
     Profiler,
