+++ tests/tests_pytorch/profilers/test_xla_profiler.py
@@ -18,7 +18,7 @@ import pytest
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.demos.boring_classes import BoringModel
-from pytorch_lightning.profiler import XLAProfiler
+from pytorch_lightning.profilers import XLAProfiler
 from pytorch_lightning.utilities import _TPU_AVAILABLE
 from tests_pytorch.helpers.runif import RunIf
